small housekeeping on one class per file
